Split up meta-request handling
