+++ internal/authorizer/meta.go
@@ -1,0 +1,12 @@
+package authorizer
+
+import "strings"
+
+const (
+	pathLabels = "/loki/api/v1/labels"
+)
+
+func isMetaRequest(path string) bool {
+	return path == pathLabels ||
+		(strings.HasPrefix(path, "/loki/api/v1/label/") && strings.HasSuffix(path, "/values"))
+}
