+++ internal/authorizer/meta_test.go
@@ -1,0 +1,30 @@
+package authorizer
+
+import "testing"
+
+func Test_isMetaRequest(t *testing.T) {
+	tests := []struct {
+		path string
+		want bool
+	}{
+		{
+			path: "/loki/api/v1/labels",
+			want: true,
+		},
+		{
+			path: "/loki/api/v1/label/kubernetes_namespace_name/values",
+			want: true,
+		},
+		{
+			path: "/loki/api/v1/query_range",
+			want: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.path, func(t *testing.T) {
+			if got := isMetaRequest(tt.path); got != tt.want {
+				t.Errorf("isMetaRequest() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
